requirements.txt
aiogram==3.0.0b7
APScheduler==3.10.1
environs==9.5.0
python-dotenv==1.0.0
tinkoff-investments==0.2.0b58
